tools/fingerprint/requirements.txt
# Základ webového agenta, ktorý poskytuje lokálne HTTPS API.
Flask>=3.1.0
flask-cors>=6.0.0

# HTTP komunikácia medzi agentom a backend serverom.
requests>=2.31.0
urllib3>=2.0.0

# Kryptografia a práca so sériovým senzorom odtlačkov.
pycryptodome>=3.20.0
pyserial>=3.5
pqcrypto>=0.4.0

# Meranie výkonu a systémových štatistík počas experimentov.
psutil>=5.9.0
